third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-layout/timeline-layout-with-invalidations-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests the Timeline API instrumentation of layout events with invalidations.


Running: testLocalFrame
first layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: test://evaluations/0/timeline-layout-with-invalidations.js:18}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
second layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: test://evaluations/0/timeline-layout-with-invalidations.js:20}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]

Running: testSubframe
first layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: test://evaluations/0/timeline-layout-with-invalidations.js:27}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
second layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: test://evaluations/0/timeline-layout-with-invalidations.js:29}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
